requirements.txt
alembic==1.12.1
asgiref==3.7.2
Cerberus==1.3.5
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
greenlet==3.0.1
gunicorn==21.2.0
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
